+++ t/Screen.t
@@ -1,4 +1,5 @@
 use v6;
+#use lib '../gnome-gobject/lib', '../gnome-native/lib';
 use Test;
 
 use Gnome::N::N-GObject;
@@ -15,11 +16,6 @@ my Gnome::Gdk3::Screen $s;
 #-------------------------------------------------------------------------------
 subtest 'ISA test', {
   throws-like
-    { $s .= new; },
-    X::Gnome, "No options used",
-    :message('No options used to create or set the native widget');
-
-  throws-like
     { $s .= new( :find, :search); },
     X::Gnome, "Wrong options used",
     :message(
@@ -28,13 +24,13 @@ subtest 'ISA test', {
           [(find||search) ',']+/
     );
 
-  $s .= new(:default);
-  isa-ok $s, Gnome::Gdk3::Screen, '.new(:default)';
+  $s .= new;
+  isa-ok $s, Gnome::Gdk3::Screen, '.new';
 }
 
 #-------------------------------------------------------------------------------
 subtest 'Manipulations', {
-  my Gnome::Gdk3::Display $display .= new(:widget($s.get-display));
+  my Gnome::Gdk3::Display $display .= new(:native-object($s.get-display));
   my Str $display-name = $display.get-name();
   like $display-name, /\: \d+/,
        '.get-display(): display name has proper format: ' ~ $display-name;